backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
erc7730>=0.3.0
python-dotenv>=0.20.0
pydantic>=2.0.0
requests>=2.26.0
google-genai>=1.0.0
mangum>=0.17.0
# python==3.12.0
boto3>=1.35.0
cryptography>=42.0.0
eth-keys>=0.4.0
eth-utils>=4.1.0